refactor: SelectSdkSummary into it's own file
