+++ frontend/src/component/onboarding/dialog/NewConnectSdkDialog/NewConnectSdkDialog.tsx
@@ -12,7 +12,8 @@ import { NewConnectSdkDialogAside } from './NewConnectSdkDialogAside';
 import { useState } from 'react';
 import { ConnectSdkDialogStep } from './ConnectSdkDialogStep';
 import type { Sdk } from '../sharedTypes';
-import { SelectSdk, SelectSdkSummary } from './SelectSdk';
+import { SelectSdk } from './SelectSdk/SelectSdk';
+import { SelectSdkSummary } from './SelectSdk/SelectSdkSummary';
 import { GenerateApiKey } from './GenerateApiKey/GenerateApiKey';
 import { GenerateApiKeySummary } from './GenerateApiKey/GenerateApiKeySummary';
 import { ConfigureSdk } from './ConfigureSdk';
