+++ frontend/src/component/onboarding/dialog/NewConnectSdkDialog/SelectSdk/SdkLogo.tsx
@@ -1,0 +1,6 @@
+import { Avatar, styled } from '@mui/material';
+
+export const SdkLogo = styled(Avatar)(({ theme }) => ({
+    width: theme.spacing(3),
+    height: theme.spacing(3),
+}));
